fix(ids): ids are integers now

+++ src/db/subscriptionPlan.js
@@ -6,7 +6,7 @@ export default (sequelize) => {
   SubscriptionPlan.init(
     {
       id: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         primaryKey: true,
         autoIncrement: true
       },
